requirements.txt
python-pptx
python-docx
openai
requests
